refactor(home) Regorganización de funciones puras
